backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
PyMySQL==1.1.1
mysql-connector-python==8.2.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
Werkzeug==3.0.1
itsdangerous==2.1.2
PyJWT==2.8.0